requirements.txt
pytest-html
pytest-xdist
pytest-ordering
pytest-rerunfailures
allure-pytest

requests~=2.31.0
pytest~=7.4.0
Flask~=2.3.2
PyYAML~=6.0